+++ v2/file_writer_html_rendering_mode_test.go
@@ -496,4 +496,3 @@ func TestFileWriterHTMLRendering_ConsecutiveFragments(t *testing.T) {
 		}
 	}
 }
-